Move Spock/Groovy tests to their own folder
This makes intelliJ compile and run the test.  It worked before. I don't understand how.    By before I mean on my old laptop and a year ago.   It does not work on my old laptop either.   I like the two test mixed together.  But this is the Maven default.  We are stuck with it until I understand it better.